refactor(wordpress): own WP Codebox runtime selection
